+++ datacube/drivers/postgres/_connections.py
@@ -17,14 +17,18 @@ import logging
 import os
 import re
 from contextlib import contextmanager
-from typing import Optional
+from time import clock_gettime, CLOCK_REALTIME
+from typing import Callable, Optional, Union
 
-from sqlalchemy import create_engine, text
+from sqlalchemy import event, create_engine, text
+from sqlalchemy.engine import Engine
 from sqlalchemy.engine.url import URL as EngineUrl
 
 import datacube
 from datacube.index.exceptions import IndexSetupError
 from datacube.utils import jsonify_document
+from datacube.utils.aws import obtain_new_iam_auth_token
+
 from . import _api
 from . import _core
 
@@ -40,6 +44,8 @@ except (ImportError, KeyError):
     # No default on Windows and some other systems
     DEFAULT_DB_USER = None
 DEFAULT_DB_PORT = 5432
+DEFAULT_IAM_AUTH = False
+DEFAULT_IAM_TIMEOUT = 600
 
 
 class PostgresDb(object):
@@ -74,12 +80,18 @@ class PostgresDb(object):
             config.get('db_port', DEFAULT_DB_PORT),
             application_name=app_name,
             validate=validate_connection,
-            pool_timeout=int(config.get('db_connection_timeout', 60))
+            iam_rds_auth=bool(config.get("db_iam_authentication", DEFAULT_IAM_AUTH)),
+            iam_rds_timeout=int(config.get("db_iam_timeout", DEFAULT_IAM_TIMEOUT)),
+            pool_timeout=int(config.get('db_connection_timeout', 60)),
+            # pass config?
         )
 
     @classmethod
     def create(cls, hostname, database, username=None, password=None, port=None,
-               application_name=None, validate=True, pool_timeout=60):
+               application_name=None, validate=True,
+               iam_rds_auth=False, iam_rds_timeout=600,
+               # pass config?
+               pool_timeout=60):
         mk_url = getattr(EngineUrl, 'create', EngineUrl)
         engine = cls._create_engine(
             mk_url(
@@ -88,6 +100,8 @@ class PostgresDb(object):
                 username=username, password=password,
             ),
             application_name=application_name,
+            iam_rds_auth=iam_rds_auth,
+            iam_rds_timeout=iam_rds_timeout,
             pool_timeout=pool_timeout)
         if validate:
             if not _core.database_exists(engine):
@@ -104,8 +118,8 @@ class PostgresDb(object):
         return PostgresDb(engine)
 
     @staticmethod
-    def _create_engine(url, application_name=None, pool_timeout=60):
-        return create_engine(
+    def _create_engine(url, application_name=None, iam_rds_auth=False, iam_rds_timeout=600, pool_timeout=60):
+        engine = create_engine(
             url,
             echo=False,
             echo_pool=False,
@@ -113,7 +127,6 @@ class PostgresDb(object):
             # 'AUTOCOMMIT' here means READ-COMMITTED isolation level with autocommit on.
             # When a transaction is needed we will do an explicit begin/commit.
             isolation_level='AUTOCOMMIT',
-
             json_serializer=_to_json,
             # If a connection is idle for this many seconds, SQLAlchemy will renew it rather
             # than assuming it's still open. Allows servers to close idle connections without clients
@@ -121,6 +134,11 @@ class PostgresDb(object):
             pool_recycle=pool_timeout,
             connect_args={'application_name': application_name}
         )
+
+        if iam_rds_auth:
+            handle_dynamic_token_authentication(engine, obtain_new_iam_auth_token, timeout=iam_rds_timeout, url=url)
+
+        return engine
 
     @property
     def url(self) -> EngineUrl:
@@ -242,6 +260,23 @@ class PostgresDb(object):
 
     def __repr__(self):
         return "PostgresDb<engine={!r}>".format(self._engine)
+
+
+def handle_dynamic_token_authentication(engine: Engine,
+                                        new_token: Callable[..., str],
+                                        timeout: Union[float, int] = 600,
+                                        **kwargs) -> None:
+    last_token = [None]
+    last_token_time = [0.0]
+
+    @event.listens_for(engine, "do_connect")
+    def override_new_connection(dialect, conn_rec, cargs, cparams):
+        # Handle IAM authentication
+        now = clock_gettime(CLOCK_REALTIME)
+        if now - last_token_time[0] > timeout:
+            last_token[0] = new_token(**kwargs)
+            last_token_time[0] = now
+        cparams["password"] = last_token[0]
 
 
 def _to_json(o):